+++ ReleaseNotes.rst
@@ -7,6 +7,9 @@ Development version @ trunk
 - The ``f50vals`` array in the ``SensitivityCube`` class has been
   replaced by an array of the 1 sigma noise called ``sigmas``
 
+- When initialising a ``SensitivityCube`` an nsigma parameter
+  can be passed that converts the input to signmas via input/nsigma
+
 - The old ``return_completeness`` method has been replaced with
   a new method where you must specify the detection 
   significance S/N cut as the final parameter. The old
@@ -16,3 +19,6 @@ Development version @ trunk
 - You may now specify a conversion between signal divided by
   noise and detection significance using the ``conversion_poly``
   parameter when initialising a ``SensitivityCube``
+
+- The alphas array can be 3D, in which case an alpha can be specified for
+each location